requirements.txt
psycopg2==2.5.3
PIL==1.1.7

# django-cms depends
django-cms==3.0.5
django==1.6.6
django-classy-tags==0.5.1
South==1.0
html5lib==1.0b3
django-mptt==0.6.1
django-sekizai==0.7
djangocms-admin-style==0.2.2
six==1.7.3

djangocms-text-ckeditor==2.2.0
djangocms-picture==0.0.4
djangocms-file==0.0.2
djangocms-flash==0.0.3
djangocms-link==1.4.0

# other
django-haystack==2.2.0
django-publisher==0.0.3
django-reversion==1.8.2
django-tinymce==1.5.2

# tinymce
# whoosh

# argparse==1.2.1
# distribute==0.6.24
# psycopg2==2.5.3
# wsgiref==0.1.2
